LibWeb: Bundle abspos layout inputs into a replayable struct
layout_absolutely_positioned_element() receives everything it consumes
from outside the element's subtree as two plain values: the containing
block info and the fully resolved static position rect. Partial
relayout will need to save those values during a full layout pass and
replay them later to re-run absolutely positioned layout without
re-running the ancestor formatting context, which requires them to be
one persistable unit.

Bundle them into an AbsposLayoutInputs struct, and move the types it
is built from (Alignment, AbsposAxisMode, AbsposContainingBlockInfo,
StaticPositionRect) to a standalone header so LayoutState can hold the
struct without including FormattingContext.h. No functional change.
