+++ Libraries/LibWeb/Layout/LayoutState.h
@@ -12,6 +12,7 @@
 #include <AK/kmalloc.h>
 #include <LibGfx/Path.h>
 #include <LibGfx/Point.h>
+#include <LibWeb/Layout/AbsposLayoutInputs.h>
 #include <LibWeb/Layout/Box.h>
 #include <LibWeb/Layout/LineBox.h>
 #include <LibWeb/Painting/Paintable.h>
@@ -27,35 +28,6 @@ enum class SizeConstraint {
 
 class AvailableSize;
 class AvailableSpace;
-
-// https://www.w3.org/TR/css-position-3/#static-position-rectangle
-struct StaticPositionRect {
-    enum class Alignment {
-        Start,
-        Center,
-        End,
-    };
-
-    CSSPixelRect rect;
-    Alignment horizontal_alignment { Alignment::Start };
-    Alignment vertical_alignment { Alignment::Start };
-
-    CSSPixelPoint aligned_position_for_box_with_size(CSSPixelSize const& size) const
-    {
-        CSSPixelPoint position = rect.location();
-        if (horizontal_alignment == Alignment::Center)
-            position.set_x(position.x() + (rect.width() - size.width()) / 2);
-        else if (horizontal_alignment == Alignment::End)
-            position.set_x(position.x() + rect.width() - size.width());
-
-        if (vertical_alignment == Alignment::Center)
-            position.set_y(position.y() + (rect.height() - size.height()) / 2);
-        else if (vertical_alignment == Alignment::End)
-            position.set_y(position.y() + rect.height() - size.height());
-
-        return position;
-    }
-};
 
 // Sparse, index-based container using two-level page tables.
 // Layout state is throwaway — rebuilt on every layout pass — so a
